Improve API specs; some impls

+++ controllers/User.js
@@ -4,7 +4,16 @@ var utils = require('../utils/writer.js');
 var User = require('../service/UserService');
 var session = require('../utils/SessionManager.js');
 
+
 module.exports.userAddAddressPOST = function userAddAddressPOST (req, res, next) {
+
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
     var addressStreetLine1 = req.swagger.params['addressStreetLine1'].value;
     var city = req.swagger.params['city'].value;
     var zip_code = req.swagger.params['zip_code'].value;
@@ -21,24 +30,34 @@ module.exports.userAddAddressPOST = function userAddAddressPOST (req, res, next)
 };
 
 module.exports.userDeletePOST = function userDeletePOST (req, res, next) {
-    try{
-        let userId = session.getUserId();
 
-        User.userDeletePOST()
-            .then(function (response) {
-                utils.writeJson(res, {message: "Succesful operation!"});
-            })
-            .catch(function (response) {
-                utils.writeJson(res, {message: "Errors performing the operation!"}, 500);
-            });
-    } catch (e){
-        utils.writeJson(res, {error: "You were not logged in!"}, 400);
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
     }
 
+    let userId = session.getUserId();
+
+    User.userDeletePOST()
+        .then(function (response) {
+            utils.writeJson(res, {message: "Succesful operation!"});
+        })
+        .catch(function (response) {
+            utils.writeJson(res, {message: "Errors performing the operation!"}, 500);
+        });
 };
 
 module.exports.userGetDetailsGET = function userGetDetailsGET (req, res, next) {
-    User.userGetDetailsGET()
+
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
+
+    User.userGetDetailsGET(userId)
         .then(function (response) {
             utils.writeJson(res, response);
         })
@@ -46,6 +65,31 @@ module.exports.userGetDetailsGET = function userGetDetailsGET (req, res, next) {
             utils.writeJson(res, response);
         });
 };
+
+module.exports.userModifyAddressPUT = function userModifyAddressPUT (req, res, next) {
+
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
+    var addressStreetLine1 = req.swagger.params['addressStreetLine1'].value;
+    var addressStreetLine2 = req.swagger.params['addressStreetLine2'].value;
+    var city = req.swagger.params['city'].value;
+    var zip_code = req.swagger.params['zip_code'].value;
+    var province = req.swagger.params['province'].value;
+    var country = req.swagger.params['country'].value;
+    User.userModifyAddressPUT(userId, addressStreetLine1,addressStreetLine2,city,zip_code,province,country)
+        .then(function (response) {
+            utils.writeJson(res, response);
+        })
+        .catch(function (response) {
+            utils.writeJson(res, response);
+        });
+};
+
 
 module.exports.userLoginPOST = function userLoginPOST (req, res, next) {
 
@@ -109,7 +153,16 @@ module.exports.userModifyPUT = function userModifyPUT (req, res, next) {
     var surname = req.swagger.params['surname'].value;
     var birthDate = req.swagger.params['birthDate'].value;
 
-    User.userModifyPUT(body)
+    if(!(username &&
+         password &&
+         email &&
+         firstName &&
+         surname &&
+         birthDate)){
+        utils.writeJson(res, {message: "No parameter set!"}, 400);
+    }
+
+    User.userModifyPUT(username, password, email, firstName, surname, birthDate)
         .then(function (response) {
             utils.writeJson(res, response);
         })
